api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
requests==2.32.3
gunicorn==23.0.0
python-multipart==0.0.12
psycopg2-binary==2.9.9
